Moved CqlBinaryRecordReader into CqlBinaryInputFormat
