+++ hadoop/CqlBinaryInputFormat.java
@@ -18,13 +18,16 @@
 
 package grakn.core.hadoop;
 
+import com.datastax.oss.driver.api.core.cql.Row;
 import com.datastax.oss.driver.internal.core.auth.PlainTextAuthProvider;
 import grakn.core.graph.diskstorage.Entry;
 import grakn.core.graph.diskstorage.StaticBuffer;
 import grakn.core.graph.diskstorage.configuration.BasicConfiguration;
 import grakn.core.graph.diskstorage.cql.CQLConfigOptions;
+import grakn.core.graph.diskstorage.cql.CQLKeyColumnValueStore;
 import grakn.core.graph.diskstorage.keycolumnvalue.SliceQuery;
 import grakn.core.graph.diskstorage.util.StaticArrayBuffer;
+import grakn.core.graph.diskstorage.util.StaticArrayEntry;
 import grakn.core.graph.graphdb.configuration.GraphDatabaseConfiguration;
 import org.apache.cassandra.hadoop.ConfigHelper;
 import org.apache.cassandra.thrift.SlicePredicate;
@@ -41,6 +44,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class CqlBinaryInputFormat extends InputFormat<StaticBuffer, Iterable<Entry>> implements HadoopPoolsConfigurable {
@@ -66,7 +70,7 @@ public class CqlBinaryInputFormat extends InputFormat<StaticBuffer, Iterable<Ent
 
     @Override
     public RecordReader<StaticBuffer, Iterable<Entry>> createRecordReader(InputSplit inputSplit, TaskAttemptContext taskAttemptContext) {
-        GraknCqlRecordReader recordReader = (GraknCqlRecordReader) cqlInputFormat.createRecordReader(inputSplit, taskAttemptContext);
+        GraknRecordReader recordReader = (GraknRecordReader) cqlInputFormat.createRecordReader(inputSplit, taskAttemptContext);
         return new CqlBinaryRecordReader(recordReader);
     }
 
@@ -115,5 +119,96 @@ public class CqlBinaryInputFormat extends InputFormat<StaticBuffer, Iterable<Ent
     @Override
     public Configuration getConf() {
         return hadoopConf;
+    }
+
+    public static class CqlBinaryRecordReader extends RecordReader<StaticBuffer, Iterable<Entry>> {
+        private KV currentKV;
+        private KV incompleteKV;
+
+        private final GraknRecordReader reader;
+
+        CqlBinaryRecordReader(GraknRecordReader reader) {
+            this.reader = reader;
+        }
+
+        @Override
+        public void initialize(InputSplit inputSplit, TaskAttemptContext taskAttemptContext) throws IOException {
+            reader.initialize(inputSplit, taskAttemptContext);
+        }
+
+        @Override
+        public boolean nextKeyValue() throws IOException {
+            return null != (currentKV = completeNextKV());
+        }
+
+        private KV completeNextKV() throws IOException {
+            KV completedKV = null;
+            boolean hasNext;
+            do {
+                hasNext = reader.nextKeyValue();
+
+                if (!hasNext) {
+                    completedKV = incompleteKV;
+                    incompleteKV = null;
+                } else {
+                    Row row = reader.getCurrentValue();
+                    StaticArrayBuffer key = StaticArrayBuffer.of(row.getBytesUnsafe(CQLKeyColumnValueStore.KEY_COLUMN_NAME));
+                    StaticBuffer column1 = StaticArrayBuffer.of(row.getBytesUnsafe(CQLKeyColumnValueStore.COLUMN_COLUMN_NAME));
+                    StaticBuffer value = StaticArrayBuffer.of(row.getBytesUnsafe(CQLKeyColumnValueStore.VALUE_COLUMN_NAME));
+                    Entry entry = StaticArrayEntry.of(column1, value);
+
+                    if (null == incompleteKV) {
+                        // Initialization; this should happen just once in an instance's lifetime
+                        incompleteKV = new KV(key);
+                    } else if (!incompleteKV.key.equals(key)) {
+                        // The underlying Cassandra reader has just changed to a key we haven't seen yet
+                        // This implies that there will be no more entries for the prior key
+                        completedKV = incompleteKV;
+                        incompleteKV = new KV(key);
+                    }
+
+                    incompleteKV.addEntry(entry);
+                }
+                /* Loop ends when either
+                 * A) the cassandra reader ran out of data
+                 * or
+                 * B) the cassandra reader switched keys, thereby completing a KV */
+            } while (hasNext && null == completedKV);
+
+            return completedKV;
+        }
+
+        @Override
+        public StaticBuffer getCurrentKey() {
+            return currentKV.key;
+        }
+
+        @Override
+        public Iterable<Entry> getCurrentValue() {
+            return currentKV.entries;
+        }
+
+        @Override
+        public void close() {
+            reader.close();
+        }
+
+        @Override
+        public float getProgress() {
+            return reader.getProgress();
+        }
+
+        private static class KV {
+            private final StaticArrayBuffer key;
+            private final ArrayList<Entry> entries = new ArrayList<>();
+
+            KV(StaticArrayBuffer key) {
+                this.key = key;
+            }
+
+            void addEntry(Entry toAdd) {
+                entries.add(toAdd);
+            }
+        }
     }
 }
